requirements.txt
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
importlib-metadata==2.0.0
inflect==4.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.6
python-dotenv==0.15.0
sqlacodegen==2.3.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
zipp==3.4.0
